feat(model): add Address, CategoryProvider, Contact, and CategoryService classes; update Provider and Service classes to use Long for IDs

+++ src/main/java/com/devtoys/queenplatform/model/Service.java
@@ -2,11 +2,11 @@ package com.devtoys.queenplatform.model;
 
 public class Service {
 
-  private Integer id;
+  private Long id;
   private String name;
   private String description;
 
-  public Service(Integer id, String name, String description) {
+  public Service(Long id, String name, String description) {
     this.id = id;
     this.name = name;
     this.description = description;
